proto: change int32 to uint32 (#4557)
* proto: change int32 to uint32

- address comments post merge from: https://github.com/tendermint/tendermint/pull/4332

proto goes down to int32/uint32 it does not provide a smaller size

* add contract to comment

* change to assertValues

* make funcs take uint32

* fix linting

* change equal to equalValues

+++ scripts/json2wal/main.go
@@ -49,7 +49,7 @@ func main() {
 	// the length of tendermint/wal/MsgInfo in the wal.json may exceed the defaultBufSize(4096) of bufio
 	// because of the byte array in BlockPart
 	// leading to unmarshal error: unexpected end of JSON input
-	br := bufio.NewReaderSize(f, 2*types.BlockPartSizeBytes)
+	br := bufio.NewReaderSize(f, int(2*types.BlockPartSizeBytes))
 	dec := cs.NewWALEncoder(walFile)
 
 	for {
